Moved Magic and MythicMobs into separate modules too.
