+++ plugin_compatibility/itemsadder/src/main/java/me/wolfyscript/utilities/compatibility/plugins/ItemsAdderImpl.java
@@ -35,7 +35,6 @@ public class ItemsAdderImpl extends PluginIntegration {
     @Override
     public void init() {
         core.registerAPIReference(new ItemsAdderRef.Parser());
-
     }
 
     @Override
