+++ plugin_compatibility/mmoitems/pom.xml
@@ -31,6 +31,12 @@
 
     <dependencies>
         <dependency>
+            <groupId>io.lumine</groupId>
+            <artifactId>MythicLib</artifactId>
+            <version>1.1.5</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
             <groupId>net.Indyuce</groupId>
             <artifactId>mmoitems</artifactId>
             <version>6.6.2</version>
