+++ plugin_compatibility/plugin-comp-artifact/pom.xml
@@ -42,6 +42,16 @@
         </dependency>
         <dependency>
             <groupId>com.wolfyscript.wolfyutilities</groupId>
+            <artifactId>magic</artifactId>
+            <version>1.8-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.wolfyscript.wolfyutilities</groupId>
+            <artifactId>mythicmobs</artifactId>
+            <version>1.8-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.wolfyscript.wolfyutilities</groupId>
             <artifactId>mmoitems</artifactId>
             <version>1.8-SNAPSHOT</version>
         </dependency>
